Brushes and tess into new modules
